Fix #4: deobfuscate SRG releases
When we use other mods' MCP releases, it breaks things if we're not on the
same MCP version. Use their SRG releases instead, and let deobfCompile
automagically make everything work. IC2 was using MCP because we told it to,
and TCon was using MCP because that particular version seems to have had a
problem reobfuscating on their end.