Adds more sanity checks to the CPacketSetSlot handler, #24
The item may only be placed in a phantom slot, As well as confirms the slot actually exists, Before inserting the item. The packet can still be modified, However it will only be able to insert items into phantom slots inside the intended GUI, So its potential to spawn actual non-phantom items should theoretically be eliminated.